refactor: transaction, cache refactorings
